split loss funcs into their own .py file
