+++ test.py
@@ -2,6 +2,7 @@ import unittest
 from KNN import KNN
 import pandas as pd
 from process_data import Data
+from loss_functions import LF
 
 # Source to understand how to test in python: https://pymbook.readthedocs.io/en/latest/testing.html and https://docs.python.org/2/library/unittest.html
 class Test(unittest.TestCase):
@@ -47,12 +48,14 @@ class Test(unittest.TestCase):
 
     def test_zero_one_loss(self):
         knn = KNN()
+        lf = LF()
         data = Data()
         data_temp = pd.read_csv(r'data/abalone.data', header=None)
         data_set = data_temp.loc[:1000][:]  # get first 100 rows of data_set
         k_val = 5
         name = 'abalone'  # used in KNN, needed here
-        self.assertIsNotNone(knn.zero_one_loss(data_set, k_val, name, data))
+        #cond_data = knn.condense_data(data_set, k_val, name, data)
+        self.assertIsNotNone(lf.zero_one_loss(data_set, k_val, name, data))
 
 # Source to understand how to test in python: https://pymbook.readthedocs.io/en/latest/testing.html and https://docs.python.org/2/library/unittest.html
 if __name__ == '__main__':
